requirements.txt
appdirs==1.4.4
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.6
django-allauth==0.41.0
django-channels==0.7.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-forms-bootstrap==3.1.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
txaio==20.4.1
urllib3==1.25.9
virtualenv==20.0.21
whitenoise==5.1.0
zope.interface==5.1.0
